app/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.5.0
python-multipart==0.0.7

# Database
psycopg2-binary==2.9.9
pgvector==0.2.5

# LLM & Embeddings
openai==1.12.0

# Data Processing (crawler, chunking)
beautifulsoup4==4.12.3
pypdf==3.17.0
tiktoken==0.5.2
tenacity==8.2.3
numpy==1.26.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
